requirements.txt
ccxt
etherscan-python
flask
web3
gunicorn